requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
transformers==4.41.2
torch==2.3.0
Pillow==10.3.0
omegaconf==2.3.0
requests==2.32.3